Add flow shards header. Improve wire translation. Translate Sub
